edit size_t

+++ kozakova.ann/T2/Delimiters.cpp
@@ -11,7 +11,7 @@ std::istream& kozakova::operator>>(std::istream& in, DelimiterString&& exp)
   }
   std::string s = "";
   char c = 0;
-  for (int i = 0; i < exp.expected.size(); ++i)
+  for (size_t i = 0; i < exp.expected.size(); ++i)
   {
     in >> c;
     s += c;
